requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
scipy>=1.10
matplotlib>=3.7
joblib>=1.3
tqdm>=4.66

lifelines>=0.28
great-expectations>=0.18
mlflow>=2.10
streamlit>=1.30

python-pptx>=0.6.23
Pillow>=10.0.0